% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Import FARS ("Fatality Analysis Reporting System") data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a valid filepath to data stored in comma separated value form}
}
\value{
This function returns a data frame containing FARS data
}
\description{
This is a function for importing a single FARS data file into R for analysis.
}
\note{
The function will stop if the \code{filename} parameter does not refer to a
   valid file.
   If the file is not in CSV form, the function will attempt to parse it anyway,
   but will return parsing failure warnings. The function should not be used for
   non-CSV files, as it will likely not properly import the data.
}
\examples{
\dontrun{fars_read("data/accident_2013.csv.bz2")}

}
