% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read year and month columns of multiple years worth of FARS data into R}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of integers representing the year values to be read. Each value
must correspond to a year with a CSV file in the working directory that can be read.}
}
\value{
a list of data frames, each containing the year and month columns of the raw data
   for each year in the input
}
\description{
This function reads the month and year columns of multiple single-year FARS data files
   into R for analysis. It is mostly useful as an input of the \code{fars_summarize_years}
   function.
}
\note{
This function will return warnings if any years are input that do not correspond
   to CSV files saved with the correct filename syntax (\code{accident_(year).csv.bz2}) in
   the working directory. Instead of a data frame containing data from the file, it will
   input a \code{NULL} value for the corresponding list element and generate the following
   warning: \code{In value[[nL]](cond) : invalid year: YYYY}
}
\examples{
\dontrun{fars_read_years(c(2013, 2014, 2015))}
\dontrun{fars_read_years(2013)}

}
