% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the number of FARS incidents in each month of several years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of integers representing the year values to be read. Each value
must correspond to a year with a CSV file in the working directory that can be read.}
}
\value{
This function returns a data frame, with rows representing calendar months and
   columns for each year in the input, with the total number of FARS entries in each month
   for the given year.
}
\description{
This function produces a summary data frame containing the number of FARS
   entries in each month of a single year or for each of several years.
}
\note{
This function will only return results columns for years that correspond to CSV
   files saved with the correct filename syntax (\code{accident_(year).csv.bz2}) in
   the current working directory. For any ears that do not match such a file, no column will
   appear in the output and the following warnign will be generated:
   \code{In value[[nL]](cond) : invalid year: YYYY}
}
\examples{
\dontrun{fars_summarize_years(c(2013,2014,2015))}
\dontrun{fars_read_years(2013)}

}
