% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Filename maker}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for which a filename is to be generated.}
}
\value{
Returns a string representing a filename.
}
\description{
This function creates a string containing a standardized filename for a CSV file containing
   a year of FARS data. It is used as an input to other functions in the package to help them
   identify FARS data files in the working directory.
}
\details{
This function takes the year input and passes it to the \code{sprintf} function, a wrapper
for the C \code{sprintf} function
}
\note{
This function requres that the year parameter be interpretable as an integer.
   For proper functioning, only integer values should be used, although numeric inputs
   will be rounded down to the nearest integer and the function will produce the requested
   output.
}
\examples{
make_filename(2013)

}
